requirements.txt
boto3
fastapi
faker
uvicorn
Jinja2
requests
python-dotenv
botocore
ipython
pydantic
google-cloud-storage
azure-appconfiguration
azure-storage-blob
azure-keyvault-secrets
azure-identity
azure-mgmt-compute
azure-mgmt-resource
azure-mgmt-network
azure-mgmt-storage
google-compute-engine
google-cloud-dns
google-cloud-appengine-admin
mysql-connector-python
SQLAlchemy
pymongo
opentelemetry-api
opentelemetry-sdk
azure-monitor-opentelemetry
